restablish mezcal page and product page
