+++ package.json
@@ -21,6 +21,7 @@
     "react": "18.2.0",
     "react-burger-menu": "^3.0.8",
     "react-dom": "18.2.0",
+    "react-icons": "^4.7.1",
     "react-tag-input-component": "^2.0.2",
     "styled-components": "^5.3.6"
   },
