+++ pages/searchBar.js
@@ -1,6 +1,7 @@
 import React from "react";
 import Head from "next/head";
 import {useRouter} from "next/router";
+import Header from "../components/Header";
 
 const search = () => {
   return (
